+++ src/main/java/com/scm/app/service/DivisionService.java
@@ -1,13 +1,14 @@
 package com.scm.app.service;
 
+import java.util.Date;
 import java.util.List;
 import java.util.Objects;
 import java.util.Optional;
 
-import com.scm.app.model.Institute;
-import com.scm.app.model.User;
+import com.scm.app.model.*;
 import com.scm.app.model.requests.PaginationRequest;
 import com.scm.app.model.response.PaginatedResponse;
+import com.scm.app.util.AuditLogger;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageRequest;
@@ -15,7 +16,6 @@ import org.springframework.data.domain.Pageable;
 import org.springframework.data.domain.Sort;
 import org.springframework.stereotype.Service;
 
-import com.scm.app.model.Division;
 import com.scm.app.repo.DivisionRepo;
 
 @Service
@@ -24,13 +24,31 @@ public class DivisionService {
 	@Autowired
 	DivisionRepo repo;
 
+	@Autowired
+	private AuditLogService auditLogService;
+
+	@Autowired
+	private AuditLogger auditLogger;
 	public Division saveDivision(Division division) {
 		Division byUserName = repo.getByName(division.getName());
 		if(byUserName!=null) {
 			throw new RuntimeException("Division is Already");
 		}
 
-		return repo.save(division);
+
+
+		Division saved = repo.save(division);
+
+		//  Audit Log for CREATE
+		auditLogger.logAction(
+				division.getAccountId(),
+				saved.getId(),
+				"Division",
+				"Create",
+				division.getCreatedBy()
+		);
+
+		return saved;
 	}
 
 
@@ -61,8 +79,25 @@ public class DivisionService {
 	}
 	
 	public boolean deleteById(Long id) {
-		repo.deleteById(id);
-		return true;
+
+		Division existing = repo.findById(id).orElse(null);
+		if (existing != null) {
+			repo.deleteById(id);
+
+			//  Audit Log for DELETE
+			auditLogger.logAction(
+					existing.getAccountId(),
+					existing.getId(),
+					"Division",
+					"Delete",
+					existing.getCreatedBy()
+			);
+			return true;
+
+		}else {
+			return false;
+		}
+
 	}
 
 	public Division updateInstitute(Division division) {
@@ -71,6 +106,18 @@ public class DivisionService {
 			throw new RuntimeException("Division is Already");
 		}
 
-		return repo.save(division);
+		Division updated = repo.save(division);
+
+		//  Audit Log for UPDATE
+		auditLogger.logAction(
+				division.getAccountId(),
+				updated.getId(),
+				"Division",
+				"Edit",
+				division.getCreatedBy()
+		);
+
+		return updated;
+
 	}
 }
