+++ src/main/java/com/scm/app/service/SchoolBranchService.java
@@ -4,10 +4,12 @@ package com.scm.app.service;
 import java.util.List;
 import java.util.Optional;
 
+import com.scm.app.model.Course;
 import com.scm.app.model.SchoolClass;
 import com.scm.app.model.User;
 import com.scm.app.model.requests.PaginationRequest;
 import com.scm.app.model.response.PaginatedResponse;
+import com.scm.app.util.AuditLogger;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageRequest;
@@ -24,15 +26,30 @@ public class SchoolBranchService {
 	@Autowired
 	SchoolBranchRepo repo;
 
+	@Autowired
+	private AuditLogger auditLogger;
+
 	public SchoolBranch saveBranch(SchoolBranch sb) {
 		SchoolBranch byName = repo.findByNameAndAccountId(sb.getName(), sb.getAccountId());
+		SchoolBranch saved = repo.save(sb);
+
 		if (byName != null) {
 			throw new RuntimeException("SchoolBranch with this name already exists");
 		}
-		return repo.save(sb);
+
+
+		//  Audit Log for CREATE
+		auditLogger.logAction(
+				sb.getAccountId(),
+				saved.getId(),
+				"SchoolBranch",
+				"Create",
+				sb.getCreatedBy()
+		);
+		return saved;
 	}
 
-	public PaginatedResponse<SchoolBranch> getAll(PaginationRequest request, Integer accountId) {
+	public PaginatedResponse<SchoolBranch> getAll(PaginationRequest request, Long accountId) {
 		Sort sort = request.getSortDir().equalsIgnoreCase("asc") ? Sort.by(request.getSortBy()).ascending() : Sort.by(request.getSortBy()).descending();
 		Pageable pageable = PageRequest.of(request.getPage(), request.getSize(), sort);
 		Page<SchoolBranch> userPage =null;
@@ -63,13 +80,41 @@ public class SchoolBranchService {
 		if (byName != null && !byName.getId().equals(sc.getId())) {
 			throw new RuntimeException("SchoolBranch with this name already exists");
 		}
-		return repo.save(sc);
+		SchoolBranch updated = repo.save(sc);
+
+		//  Audit Log for UPDATE
+		auditLogger.logAction(
+				sc.getAccountId(),
+				updated.getId(),
+				"Course",
+				"Edit",
+				sc.getCreatedBy()
+		);
+
+		return updated;
+
 	}
 
 
 	public boolean deleteById(Long id) {
-		repo.deleteById(id);
-		return true;
+		SchoolBranch existing = repo.findById(id).orElse(null);
+		if (existing != null) {
+			repo.deleteById(id);
+			//  Audit Log for DELETE
+			auditLogger.logAction(
+					existing.getAccountId(),
+					existing.getId(),
+					"SchoolBranch",
+					"Delete",
+					existing.getCreatedBy()
+			);
+			return true;
+
+		}else {
+			return false;
+		}
+
+
 	}
 	
 }
