+++ src/main/java/com/scm/app/service/SchoolClassService.java
@@ -3,10 +3,12 @@ package com.scm.app.service;
 import java.util.List;
 import java.util.Optional;
 
+import com.scm.app.model.Course;
 import com.scm.app.model.Subject;
 import com.scm.app.model.User;
 import com.scm.app.model.requests.PaginationRequest;
 import com.scm.app.model.response.PaginatedResponse;
+import com.scm.app.util.AuditLogger;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageRequest;
@@ -23,12 +25,25 @@ public class SchoolClassService {
 	@Autowired
 	SchoolClassRepo repo;
 
+	@Autowired
+	private AuditLogger auditLogger;
+
 	public SchoolClass saveClass(SchoolClass sc) {
 		SchoolClass byName = repo.findByNameAndAccountId(sc.getName(), sc.getAccountId());
 		if (byName != null) {
 			throw new RuntimeException("SchoolClass with this name already exists");
 		}
-		return repo.save(sc);
+		SchoolClass saved = repo.save(sc);
+
+		//  Audit Log for CREATE
+		auditLogger.logAction(
+				sc.getAccountId(),
+				saved.getId(),
+				"Course",
+				"Create",
+				sc.getCreatedBy()
+		);
+		return saved;
 	}
 
 	public SchoolClass updateClass(SchoolClass sc) {
@@ -36,11 +51,23 @@ public class SchoolClassService {
 		if (byName != null && !byName.getId().equals(sc.getId())) {
 			throw new RuntimeException("SchoolClass with this name already exists");
 		}
-		return repo.save(sc);
+		SchoolClass updated = repo.save(sc);
+
+		//  Audit Log for UPDATE
+		auditLogger.logAction(
+				sc.getAccountId(),
+				updated.getId(),
+				"Course",
+				"Edit",
+				sc.getCreatedBy()
+		);
+
+		return updated;
+
 	}
 
 
-	public PaginatedResponse<SchoolClass> getAll(PaginationRequest request, Integer accountId) {
+	public PaginatedResponse<SchoolClass> getAll(PaginationRequest request, Long accountId) {
 		Sort sort = request.getSortDir().equalsIgnoreCase("asc") ? Sort.by(request.getSortBy()).ascending() : Sort.by(request.getSortBy()).descending();
 		Pageable pageable = PageRequest.of(request.getPage(), request.getSize(), sort);
 
@@ -68,8 +95,24 @@ public class SchoolClassService {
 	}
 	
 	public boolean deleteById(Long id) {
-		repo.deleteById(id);
-		return true;
+		SchoolClass existing = repo.findById(id).orElse(null);
+		if (existing != null) {
+			repo.deleteById(id);
+
+			//  Audit Log for DELETE
+			auditLogger.logAction(
+					existing.getAccountId(),
+					existing.getId(),
+					"SchoolClass ",
+					"Delete",
+					existing.getCreatedBy()
+			);
+			return true;
+
+		}else {
+			return false;
+		}
+
 	}
 
 }
