+++ src/main/java/com/scm/app/service/StudentService.java
@@ -3,6 +3,9 @@ package com.scm.app.service;
 import java.util.List;
 import java.util.Optional;
 
+import com.scm.app.model.Assignment;
+import com.scm.app.model.Course;
+import com.scm.app.util.AuditLogger;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
@@ -15,8 +18,23 @@ public class StudentService {
 	@Autowired
 	StudentRepo repo;
 
+	@Autowired
+	private AuditLogger auditLogger;
+
 	public Student saveStudent(Student std) {
-		return repo.save(std);
+		Student saved = repo.save(std);
+
+		auditLogger.logAction(
+				std.getAccountId(),
+				saved.getRollno(),
+				"Student",
+				"Create",
+				std.getCreatedBy()
+		);
+
+
+		return saved;
+
 	}
 
 	public List<Student> getAll() {
@@ -29,13 +47,39 @@ public class StudentService {
 	}
 
 	public Student saveInstitute(Student student) {
-		return repo.save(student);
+		Student saved = repo.save(student);
+		auditLogger.logAction(
+				student.getAccountId(),
+				saved.getRollno(),
+				"Student",
+				"Create",
+				student.getCreatedBy()
+		);
+
+
+		return saved;
 	}
 	
 	
 	public boolean deleteById(Long id) {
-		repo.deleteById(id);
-		return true;
+		Student existing = repo.findById(id).orElse(null);
+		if (existing != null) {
+			repo.deleteById(id);
+
+			//  Audit Log for DELETE
+			auditLogger.logAction(
+					existing.getAccountId(),
+					existing.getRollno(),
+					"Student",
+					"Delete",
+					existing.getCreatedBy()
+			);
+			return true;
+
+		}else {
+			return false;
+		}
+
 	}
 
 	public List<Student> getByDivisionIdAndClassId(Long classId, Long divisionId) {
