+++ src/main/java/com/scm/app/service/SubjectService.java
@@ -3,9 +3,11 @@ package com.scm.app.service;
 import java.util.List;
 import java.util.Optional;
 
+import com.scm.app.model.Course;
 import com.scm.app.model.User;
 import com.scm.app.model.requests.PaginationRequest;
 import com.scm.app.model.response.PaginatedResponse;
+import com.scm.app.util.AuditLogger;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageRequest;
@@ -22,12 +24,26 @@ public class SubjectService {
 	@Autowired
 	SubjectRepo repo;
 
+	@Autowired
+	private AuditLogger auditLogger;
+
 	public Subject saveSubject(Subject std) {
 		Subject byName = repo.findByNameAndAccountId(std.getName(), std.getAccountId());
 		if (byName != null) {
 			throw new RuntimeException("Subject with this name already exists");
 		}
-		return repo.save(std);
+		Subject saved = repo.save(std);
+
+		//  Audit Log for CREATE
+		auditLogger.logAction(
+				std.getAccountId(),
+				saved.getId(),
+				"Subject",
+				"Create",
+				std.getCreatedBy()
+		);
+		return saved;
+
 	}
 
 	public Subject updateSubject(Subject std) {
@@ -35,10 +51,22 @@ public class SubjectService {
 		if (byName != null && !byName.getId().equals(std.getId())) {
 			throw new RuntimeException("Subject with this name already exists");
 		}
-		return repo.save(std);
+		Subject updated = repo.save(std);
+
+		//  Audit Log for UPDATE
+		auditLogger.logAction(
+				std.getAccountId(),
+				updated.getId(),
+				"Subject",
+				"Edit",
+				std.getCreatedBy()
+		);
+
+		return updated;
+
 	}
 
-	public PaginatedResponse<Subject> getAll(PaginationRequest request, Integer accountId) {
+	public PaginatedResponse<Subject> getAll(PaginationRequest request, Long accountId) {
 
 		Sort sort = request.getSortDir().equalsIgnoreCase("asc") ? Sort.by(request.getSortBy()).ascending() : Sort.by(request.getSortBy()).descending();
 		Pageable pageable = PageRequest.of(request.getPage(), request.getSize(), sort);
@@ -66,8 +94,24 @@ public class SubjectService {
 	}
 	
 	public boolean deleteById(Long id) {
-		repo.deleteById(id);
-		return true;
+		Subject existing = repo.findById(id).orElse(null);
+		if (existing != null) {
+			repo.deleteById(id);
+
+			//  Audit Log for DELETE
+			auditLogger.logAction(
+					existing.getAccountId(),
+					existing.getId(),
+					"Subject",
+					"Delete",
+					existing.getCreatedBy()
+			);
+			return true;
+
+		}else {
+			return false;
+		}
+
 	}
 
 }
