+++ src/main/java/com/scm/app/service/TeacherService.java
@@ -2,18 +2,17 @@ package com.scm.app.service;
 
 import java.util.Optional;
 
-import com.scm.app.model.TimeTable;
-import com.scm.app.model.User;
+import com.scm.app.model.*;
 import com.scm.app.model.requests.PaginationRequest;
 import com.scm.app.model.response.PaginatedResponse;
 import com.scm.app.repo.TimeTableRepo;
+import com.scm.app.util.AuditLogger;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageRequest;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.domain.Sort;
 import org.springframework.stereotype.Service;
-import com.scm.app.model.Teacher;
 import com.scm.app.repo.TeacherRepo;
 
 @Service
@@ -22,12 +21,26 @@ public class TeacherService {
 	@Autowired
 	TeacherRepo repo;
 
+	@Autowired
+	private AuditLogger auditLogger;
 
 	public Teacher saveTeacher(Teacher std) {
-		return repo.save(std);
+		Teacher saved = repo.save(std);
+
+		auditLogger.logAction(
+				std.getAccountId(),
+				saved.getId(),
+				"Teacher",
+				"Create",
+				std.getCreatedBy()
+		);
+
+
+		return saved;
+
 	}
 
-	public PaginatedResponse<Teacher> getAll(PaginationRequest request, Integer accountId) {
+	public PaginatedResponse<Teacher> getAll(PaginationRequest request, Long accountId) {
 
 		Sort sort = request.getSortDir().equalsIgnoreCase("asc") ? Sort.by(request.getSortBy()).ascending() : Sort.by(request.getSortBy()).descending();
 		Pageable pageable = PageRequest.of(request.getPage(), request.getSize(), sort);
@@ -56,8 +69,24 @@ public class TeacherService {
 	}
 	
 	public boolean deleteById(Long id) {
-		repo.deleteById(id);
-		return true;
+		Teacher existing = repo.findById(id).orElse(null);
+		if (existing != null) {
+			repo.deleteById(id);
+
+			//  Audit Log for DELETE
+			auditLogger.logAction(
+					existing.getAccountId(),
+					existing.getId(),
+					"Teacher",
+					"Delete",
+					existing.getCreatedBy()
+			);
+			return true;
+
+		}else {
+			return false;
+		}
+
 	}
 
 }
