+++ src/main/java/com/scm/app/service/TimeTableService.java
@@ -3,9 +3,12 @@ package com.scm.app.service;
 import java.util.List;
 import java.util.Optional;
 
+import com.scm.app.model.Assignment;
+import com.scm.app.model.Course;
 import com.scm.app.model.User;
 import com.scm.app.model.requests.PaginationRequest;
 import com.scm.app.model.response.PaginatedResponse;
+import com.scm.app.util.AuditLogger;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageRequest;
@@ -22,14 +25,30 @@ public class TimeTableService {
 	@Autowired
 	TimeTableRepo repo;
 
+	@Autowired
+	private AuditLogger auditLogger;
+
 	public TimeTable saveTimeTable(TimeTable timeTable) {
-		
 		timeTable.setDayTimeTable(timeTable.getDayTimeTable());
 
-		return repo.save(timeTable);
+		TimeTable saved = repo.save(timeTable);
+
+		auditLogger.logAction(
+				timeTable.getAccountId(),
+				saved.getId(),
+				"TimeTable",
+				"Create",
+				timeTable.getCreatedBy()
+		);
+
+
+		return saved;
+		
+
+
 	}
 
-	public PaginatedResponse<TimeTable> getAll(PaginationRequest request, Integer accountId) {
+	public PaginatedResponse<TimeTable> getAll(PaginationRequest request, Long accountId) {
 
 		Sort sort = request.getSortDir().equalsIgnoreCase("asc") ? Sort.by(request.getSortBy()).ascending() : Sort.by(request.getSortBy()).descending();
 		Pageable pageable = PageRequest.of(request.getPage(), request.getSize(), sort);
@@ -57,7 +76,23 @@ public class TimeTableService {
 	}
 
 	public boolean deleteById(Long id) {
-		repo.deleteById(id);
-		return true;
+		TimeTable existing = repo.findById(id).orElse(null);
+		if (existing != null) {
+			repo.deleteById(id);
+
+			//  Audit Log for DELETE
+			auditLogger.logAction(
+					existing.getAccountId(),
+					existing.getId(),
+					"TimeTable",
+					"Delete",
+					existing.getCreatedBy()
+			);
+			return true;
+
+		}else {
+			return false;
+		}
+
 	}
 }
